Move contact verification helpers to lib
